split qlinear activation family
